move logout view under devhub
