+++ src/olympia/devhub/signals.py
@@ -1,4 +1,5 @@
 import django.dispatch
 
 
+logged_out = django.dispatch.Signal(providing_args=['request', 'response'])
 submission_done = django.dispatch.Signal()
