+++ src/olympia/lib/settings_base.py
@@ -1027,7 +1027,6 @@ MESSAGE_STORAGE = 'django.contrib.messages.storage.cookie.CookieStorage'
 
 # These should have app+locale at the start to avoid redirects
 LOGIN_URL = reverse_lazy('users.login')
-LOGOUT_URL = reverse_lazy('users.logout')
 LOGIN_REDIRECT_URL = '/'
 LOGOUT_REDIRECT_URL = '/'
 # When logging in with browser ID, a username is created automatically.
