--- src/olympia/users/signals.py
@@ -1,4 +1,0 @@
-from django.dispatch import Signal
-
-
-logged_out = Signal(providing_args=['request', 'response'])
