+++ src/olympia/users/views.py
@@ -22,7 +22,7 @@ from olympia.amo import messages
 from olympia.amo.decorators import (
     json_view, login_required, permission_required, use_primary_db)
 from olympia.amo.forms import AbuseForm
-from olympia.amo.urlresolvers import get_url_prefix, reverse
+from olympia.amo.urlresolvers import reverse
 from olympia.amo.utils import escape_all, render
 from olympia.bandwagon.models import Collection
 from olympia.browse.views import PersonasFilter
@@ -31,7 +31,6 @@ from olympia.users.models import UserNotification
 
 from . import forms, tasks
 from .models import UserProfile
-from .signals import logged_out
 from .utils import UnsubscribeCode
 
 
@@ -196,30 +195,6 @@ def login(request):
         return http.HttpResponseRedirect(request.GET['to'])
     else:
         return render(request, 'users/login.html')
-
-
-def logout(request):
-    user = request.user
-    if not user.is_anonymous:
-        log.debug(u"User (%s) logged out" % user)
-
-    if 'to' in request.GET:
-        request = _clean_next_url(request)
-
-    next_url = request.GET.get('to')
-    if not next_url:
-        next_url = settings.LOGOUT_REDIRECT_URL
-        prefixer = get_url_prefix()
-        if prefixer:
-            next_url = prefixer.fix(next_url)
-
-    response = http.HttpResponseRedirect(next_url)
-
-    logout_user(request, response)
-
-    # Fire logged out signal.
-    logged_out.send(None, request=request, response=response)
-    return response
 
 
 @user_view
